Refactored tests to use typescript
