+++ package.json
@@ -5,7 +5,7 @@
   "main": "lib/aat.js",
   "types": "lib/aat.d.ts",
   "scripts": {
-    "test": "mocha --reporter spec",
+    "test": "mocha -r ts-node/register tests/**/*.spec.ts",
     "build": "tsc",
     "format": "prettier --write \"src/**/*.ts\" \"src/**/*.js\"",
     "lint": "tslint -p tsconfig.json"
@@ -25,9 +25,12 @@
     "lib/**/*"
   ],
   "devDependencies": {
+    "@types/chai": "^4.2.7",
+    "@types/mocha": "^5.2.7",
     "chai": "^4.2.0",
     "mocha": "^6.2.2",
     "prettier": "^1.19.1",
+    "ts-node": "^8.5.4",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
     "typescript": "^3.7.2"
